Moved gyro to utility class
